Fix: templates.temperature column must be DOUBLE PRECISION, not REAL
REAL (float4) truncates precision on write, so a Go float64 like 0.7
round-trips as ~0.699999988079071 after the float4->float8 widen on
read, failing TestTemplateAgentOverridesRoundTrip's exact comparison.
model.Template.Temperature is *float64, and the existing convention
for *float64 columns in this codebase (transcript_segments.confidence,
see 20260702085938_segment_diarization_confidence.up.sql) is
DOUBLE PRECISION, not REAL. Switch the migration to match.

+++ internal/db/migrations/20260715003612_template_agent_overrides.up.sql
@@ -2,4 +2,4 @@
 ALTER TABLE templates
 ADD COLUMN system_prompt TEXT,
 ADD COLUMN model TEXT,
-ADD COLUMN temperature REAL;
+ADD COLUMN temperature DOUBLE PRECISION;
